requirements.txt
selenium
chromedriver-binary-auto
beautifulsoup4
requests
